+++ src/java/controller/BookTicketServlet.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package controller.hompage;
+package controller;
 
 import dal.MovieDAO;
 import dal.RoomDAO;
