+++ src/java/controller/MovieDetailServlet.java
@@ -2,12 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-package controller.hompage;
+package controller;
 
 import dal.MovieDAO;
-import dal.RoomDAO;
-import dal.SeatDAO;
-import dal.ShowtimingDAO;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
@@ -16,15 +13,13 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.List;
 import model.Movie;
-import model.Room;
-import model.Seat;
-import model.Showtiming;
 
 /**
  *
  * @author thanh
  */
-public class SeatServlet extends HttpServlet {
+//url : moviedetail
+public class MovieDetailServlet extends HttpServlet {
 
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
@@ -43,10 +38,10 @@ public class SeatServlet extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet SeatServlet</title>");            
+            out.println("<title>Servlet MovieDetailServlet</title>");            
             out.println("</head>");
             out.println("<body>");
-            out.println("<h1>Servlet SeatServlet at " + request.getContextPath() + "</h1>");
+            out.println("<h1>Servlet MovieDetailServlet at " + request.getContextPath() + "</h1>");
             out.println("</body>");
             out.println("</html>");
         }
@@ -61,28 +56,20 @@ public class SeatServlet extends HttpServlet {
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
+    //author: Thanh
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-       SeatDAO sdao = new SeatDAO();
-       List<Seat> listS;
-       String roomID = request.getParameter("roomID");
-       MovieDAO moviedao = new MovieDAO();
-      ShowtimingDAO showdao = new ShowtimingDAO();
-       try {
-           int roomid = Integer.parseInt(roomID);
-           RoomDAO rdao = new RoomDAO();
-           Room room = rdao.getRoomByID(roomid);
-           listS = sdao.getSeatsByCharacterName(room.getRoom_name());
-           request.setAttribute("listS", listS);
-           List<Showtiming> listshow = showdao.getShowtimingByRoomID(roomid);
-           
-           Movie movie = moviedao.getMovieWithListShowtiming(listshow);
-           request.setAttribute("movie", movie);
-       } catch(Exception e) {
-           
-       }
-       request.getRequestDispatcher("/views/seat_selection/Seat.jsp").forward(request, response);
+        MovieDAO mdao = new MovieDAO();
+        String id = request.getParameter("ID");
+        try {
+            int idMovie = Integer.parseInt(id);
+            Movie m = mdao.getMovieById(idMovie);
+            request.setAttribute("m", m);
+            request.getRequestDispatcher("/views/homepage/MovieDetail.jsp").forward(request, response);
+        } catch(NumberFormatException e) {
+            e.printStackTrace();
+        }
     }
 
     /**
