+++ web/WEB-INF/web.xml
@@ -1,4 +1,4 @@
-\<?xml version="1.0" encoding="UTF-8"?>
+<?xml version="1.0" encoding="UTF-8"?>
 <web-app version="2.5" xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
     <servlet>
         <servlet-name>SeatServlet</servlet-name>
@@ -7,10 +7,6 @@
     <servlet>
         <servlet-name>ShowtimingServlet</servlet-name>
         <servlet-class>controller.ShowtimingServlet</servlet-class>
-    </servlet>
-    <servlet>
-        <servlet-name>RoomServlet</servlet-name>
-        <servlet-class>controller.RoomServlet</servlet-class>
     </servlet>
     <servlet>
         <servlet-name>FilterServlet</servlet-name>
@@ -23,10 +19,6 @@
     <servlet-mapping>
         <servlet-name>ShowtimingServlet</servlet-name>
         <url-pattern>/showtiming</url-pattern>
-    </servlet-mapping>
-    <servlet-mapping>
-        <servlet-name>RoomServlet</servlet-name>
-        <url-pattern>/room</url-pattern>
     </servlet-mapping>
     <servlet-mapping>
         <servlet-name>FilterServlet</servlet-name>
@@ -125,9 +117,9 @@
         <servlet-name>LogoutController</servlet-name>
         <url-pattern>/logout</url-pattern>
     </servlet-mapping>
-        <servlet>
+   <servlet>
         <servlet-name>BookTicketServlet</servlet-name>
-        <servlet-class>controller.homepage.BookTicketServlet</servlet-class>
+        <servlet-class>controller.BookTicketServlet</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>BookTicketServlet</servlet-name>
